requirements.txt
telethon>=1.40.0
requests>=2.28.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.11.0